Move label methods around
