+++ ui-timeline/src/commonMain/kotlin/com/tunjid/heron/timeline/utilities/PostExt.kt
@@ -16,6 +16,50 @@
 
 package com.tunjid.heron.timeline.utilities
 
+import com.tunjid.heron.data.core.models.ContentLabelPreference
+import com.tunjid.heron.data.core.models.ContentLabelPreferences
+import com.tunjid.heron.data.core.models.Label
+import com.tunjid.heron.data.core.models.Labelers
 import com.tunjid.heron.data.core.models.Post
 
 val Post.createdAt get() = record?.createdAt ?: indexedAt
+
+fun Post.blurredMediaLabels(
+    labelers: Labelers,
+    contentPreferences: ContentLabelPreferences,
+): List<Label.Value> {
+    if (labels.isEmpty()) return emptyList()
+
+    val contentPreferenceMap = contentPreferences.associateBy(
+        keySelector = ContentLabelPreference::label,
+        valueTransform = { it.visibility }
+    )
+
+    val mediaBlurLabels = labelers.flatMapTo(mutableSetOf()) { labeler ->
+        labeler.definitions.mapNotNull { definition ->
+            when {
+                definition.adultOnly -> definition.identifier
+                definition.blurs == Label.BlurTarget.Media -> definition.identifier
+                else -> null
+            }
+        }
+    }
+    val postBlurredLabels = labels.mapNotNull { label ->
+        if (mediaBlurLabels.contains(label.value)) when (
+            val visibility = contentPreferenceMap[label.value]
+        ) {
+            null -> label.value
+            else -> label.value.takeIf { visibility.shouldBlurMedia }
+        }
+        else null
+    }
+
+    return postBlurredLabels
+}
+
+val Label.Visibility.shouldBlurMedia
+    get() = when (this) {
+        Label.Visibility.Hide -> true
+        Label.Visibility.Warn -> true
+        else -> false
+    }
